# resources/Construcao-Software-Swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Construcao-Software",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://192.168.0.8:3000",
      "variables": {}
    }
  ],
  "paths": {
    "/api/students": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get All",
        "operationId": "GetAll",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetAll"
                  },
                  "description": "",
                  "example": [
                    {
                      "_id": "618c4c7630040e8e1eeab417",
                      "name": "john",
                      "email": "potato@potate.com",
                      "birthday": "2019-08-21T00:00:00.000Z",
                      "phone": "+55(51)99455-6722"
                    }
                  ]
                },
                "example": [
                  {
                    "_id": "618c4c7630040e8e1eeab417",
                    "name": "john",
                    "email": "potato@potate.com",
                    "birthday": "2019-08-21T00:00:00.000Z",
                    "phone": "+55(51)99455-6722"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "post": {
        "tags": [
          "students"
        ],
        "summary": "Register",
        "operationId": "Register",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "example": {
                "name": "john",
                "email": "potato@ptate.com",
                "birthday": "2019-08-21",
                "phone": "+55(51)99455-6722"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV) (RFC 4918)",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RegisterMissingValue1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Invalid value",
                      "param": "birthday",
                      "location": "body"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Invalid value",
                    "param": "birthday",
                    "location": "body"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
    "/api/students/{studentid}": {
      "get": {
        "tags": [
          "students"
        ],
        "summary": "Get One",
        "operationId": "GetOne",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetOne"
                },
                "example": {
                  "_id": "618c4c7630040e8e1eeab417",
                  "name": "john",
                  "email": "potato@potate.com",
                  "birthday": "2019-08-21T00:00:00.000Z",
                  "phone": "+55(51)99455-6722"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetOneWrongFormat1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Cast to ObjectId failed for value \"drftgyvbhjn\" (type string) at path \"_id\" for model \"Students\""
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Cast to ObjectId failed for value \"drftgyvbhjn\" (type string) at path \"_id\" for model \"Students\""
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetOneWrongID1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "student not found with id=614b7defed98c3e7891ec4de"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "student not found with id=614b7defed98c3e7891ec4de"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "put": {
        "tags": [
          "students"
        ],
        "summary": "Total Update",
        "operationId": "TotalUpdate",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotalUpdateRequest"
              },
              "example": {
                "name": "johnny",
                "email": "potato@batata.com",
                "birthday": "2019-08-20",
                "phone": "+55(51)99455-6422"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV) (RFC 4918)",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TotalUpdateMissingValue1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Invalid value",
                      "param": "birthday",
                      "location": "body"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Invalid value",
                    "param": "birthday",
                    "location": "body"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "patch": {
        "tags": [
          "students"
        ],
        "summary": "Partial Update",
        "operationId": "PartialUpdate",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PartialUpdateRequest"
              },
              "example": {
                "name": "johnny",
                "email": "potato@batata.com",
                "birthday": "2019-08-20"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "422": {
            "description": "Unprocessable Entity (WebDAV) (RFC 4918)",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartialUpdateWrongType1"
                  },
                  "description": "",
                  "example": [
                    {
                      "value": "potatobatata.com",
                      "msg": "Invalid value",
                      "param": "email",
                      "location": "body"
                    }
                  ]
                },
                "example": [
                  {
                    "value": "potatobatata.com",
                    "msg": "Invalid value",
                    "param": "email",
                    "location": "body"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      },
      "delete": {
        "tags": [
          "students"
        ],
        "summary": "Remove",
        "operationId": "Remove",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "example": "{{access_token}}"
            }
          },
          {
            "name": "studentid",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {},
            "content": {}
          },
          "400": {
            "description": "Bad Request",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemoveWrongFormat1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "Cast to ObjectId failed for value \"kll\" (type string) at path \"_id\" for model \"Students\""
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "Cast to ObjectId failed for value \"kll\" (type string) at path \"_id\" for model \"Students\""
                  }
                ]
              }
            }
          },
          "404": {
            "description": "Not Found",
            "headers": {},
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemoveWrongID1"
                  },
                  "description": "",
                  "example": [
                    {
                      "msg": "student not found with id=614b7deced98c3e7891ec4de"
                    }
                  ]
                },
                "example": [
                  {
                    "msg": "student not found with id=614b7deced98c3e7891ec4de"
                  }
                ]
              }
            }
          }
        },
        "deprecated": false,
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "GetAll": {
        "title": "GetAll",
        "required": [
          "_id",
          "name",
          "email",
          "birthday",
          "phone"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "example": {
          "_id": "618c4c7630040e8e1eeab417",
          "name": "john",
          "email": "potato@potate.com",
          "birthday": "2019-08-21T00:00:00.000Z",
          "phone": "+55(51)99455-6722"
        }
      },
      "GetOne": {
        "title": "GetOne",
        "required": [
          "_id",
          "name",
          "email",
          "birthday",
          "phone"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "example": {
          "_id": "618c4c7630040e8e1eeab417",
          "name": "john",
          "email": "potato@potate.com",
          "birthday": "2019-08-21T00:00:00.000Z",
          "phone": "+55(51)99455-6722"
        }
      },
      "GetOneWrongFormat1": {
        "title": "GetOneWrongFormat1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Cast to ObjectId failed for value \"drftgyvbhjn\" (type string) at path \"_id\" for model \"Students\""
        }
      },
      "GetOneWrongID1": {
        "title": "GetOneWrongID1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "student not found with id=614b7defed98c3e7891ec4de"
        }
      },
      "RegisterRequest": {
        "title": "RegisterRequest",
        "required": [
          "name",
          "email",
          "birthday",
          "phone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "example": {
          "name": "john",
          "email": "potato@ptate.com",
          "birthday": "2019-08-21",
          "phone": "+55(51)99455-6722"
        }
      },
      "RegisterMissingValue1": {
        "title": "RegisterMissingValue1",
        "required": [
          "msg",
          "param",
          "location"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Invalid value",
          "param": "birthday",
          "location": "body"
        }
      },
...
      "TotalUpdateRequest": {
        "title": "TotalUpdateRequest",
        "required": [
          "name",
          "email",
          "birthday",
          "phone"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "example": {
          "name": "johnny",
          "email": "potato@batata.com",
          "birthday": "2019-08-20",
          "phone": "+55(51)99455-6422"
        }
      },
      "TotalUpdateMissingValue1": {
        "title": "TotalUpdateMissingValue1",
        "required": [
          "msg",
          "param",
          "location"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Invalid value",
          "param": "birthday",
          "location": "body"
        }
      },
...
      "PartialUpdateRequest": {
        "title": "PartialUpdateRequest",
        "required": [
          "name",
          "email",
          "birthday"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "birthday": {
            "type": "string"
          }
        },
        "example": {
          "name": "johnny",
          "email": "potato@batata.com",
          "birthday": "2019-08-20"
        }
      },
      "PartialUpdateWrongType1": {
        "title": "PartialUpdateWrongType1",
        "required": [
          "value",
          "msg",
          "param",
          "location"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          },
          "param": {
            "type": "string"
          },
          "location": {
            "type": "string"
          }
        },
        "example": {
          "value": "potatobatata.com",
          "msg": "Invalid value",
          "param": "email",
          "location": "body"
        }
      },
      "RemoveWrongFormat1": {
        "title": "RemoveWrongFormat1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "Cast to ObjectId failed for value \"kll\" (type string) at path \"_id\" for model \"Students\""
        }
      },
      "RemoveWrongID1": {
        "title": "RemoveWrongID1",
        "required": [
          "msg"
        ],
        "type": "object",
        "properties": {
          "msg": {
            "type": "string"
          }
        },
        "example": {
          "msg": "student not found with id=614b7deced98c3e7891ec4de"
        }
      },
...
    }
  },
  "security": [
    {}
  ],
...
}